+++ src/main/java/com/example/web/controller/admin/ClubController.java
@@ -4,16 +4,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import com.example.service.admin.ClubService;
 import com.example.vo.Club;
-import com.example.web.request.ClubModifyForm;
 
 @Controller
-@RequestMapping("/emp/admin/club")
+@RequestMapping("/emp/club")
 public class ClubController {
 
 	@Autowired
@@ -25,20 +22,5 @@ public class ClubController {
 		model.addAttribute("club", club);
 		
 		return "admin/club/club";
-	}
-	
-	@GetMapping("/modify")
-	public String modifyForm(@RequestParam(name = "no") int no, Model model) {
-		Club club = clubService.getClub();
-		model.addAttribute("club", club);
-		
-		return "admin/club/modify";
-	}
-	
-	@PostMapping("/modify")
-	public String modify(ClubModifyForm form) {
-		clubService.updateClub(form);
-		
-		return "redirect:club";
 	}
 }
