+++ src/main/java/com/example/web/controller/admin/ProgramController.java
@@ -8,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -17,13 +16,10 @@ import com.example.dto.ProgramDayDto;
 import com.example.dto.ProgramDto;
 import com.example.service.admin.EmployeeService;
 import com.example.service.admin.ProgramService;
-import com.example.vo.Employee;
-import com.example.vo.ProgramCategory;
 import com.example.vo.User;
-import com.example.web.request.ProgramForm;
 
 @Controller
-@RequestMapping("/emp/admin/program")
+@RequestMapping("/emp/program")
 public class ProgramController {
 
 	@Autowired
@@ -62,105 +58,5 @@ public class ProgramController {
 		detailInfo.put("users", users);
 		
 		return detailInfo; 
-	}
-	
-	// 직원 검색
-	@GetMapping("/searchEmployees")
-	@ResponseBody
-	public List<Employee> employeeList(@RequestParam(name = "employeeName", required = false) String employeeName) {
-		List<Employee> employeeList = employeeService.getSearchEmployees(employeeName);
-		return employeeList;
-	}
-	
-	// 프로그램 수정 페이지
-	@GetMapping("/modify")
-	public String modifyForm(@RequestParam(name="programNo") int programNo, Model model) {
-		// 프로그램 상세 정보
-		ProgramDto programDetail = programService.getProgramDetail(programNo);
-		// 프로그램 분류 목록
-		List<ProgramCategory> categories = programService.getProgramCategory();
-		// 재직중인 직원 정보 목록
-		List<Employee> employeeList = employeeService.getEmployeeByStatus();
-		// 프로그램 선택 요일 정보
-		List<ProgramDayDto> programDays = programService.getDayStatusByProgramNo(programNo);
-		
-		model.addAttribute("programDetail", programDetail);
-		model.addAttribute("categories", categories);
-		model.addAttribute("employeeList", employeeList);
-		model.addAttribute("programDays", programDays);
-		
-		return "admin/program/modify";
-	}
-	
-	// 프로그램 수정 요청
-	@PostMapping("/modify")
-	public String modify(ProgramForm form) {
-		programService.updateProgram(form);
-		
-		return "redirect:list";
-	}
-	
-	// 프로그램 등록 페이지
-	@GetMapping("/insert")
-	public String insertForm(Model model) {
-		// 프로그램 분류 목록
-		List<ProgramCategory> categories = programService.getProgramCategory();
-		// 재직중인 직원 정보 목록
-		List<Employee> employeeList = employeeService.getEmployeeByStatus();
-		
-		model.addAttribute("categories", categories);
-		model.addAttribute("employeeList", employeeList);
-		
-		return "admin/program/insert";
-	}
-	
-	// 프로그램 등록 요청
-	@PostMapping("/insert")
-	public String insert(ProgramForm form) {
-		programService.insertProgram(form);
-		return "redirect:list";
-	}
-	
-	// 프로그램 삭제
-	@GetMapping("/delete")
-	public String delete(@RequestParam(name="programNo") String programNo) {
-		programService.deleteProgram(programNo);
-		return "redirect:list";
-	}
-	
-	// 프로그램 분류 페이지
-	@GetMapping("/category")
-	public String category(Model model) {
-		// 프로그램 분류 목록
-		List<ProgramCategory> categories = programService.getProgramCategory();
-		model.addAttribute("categories",categories);
-		
-		return "admin/program/category";
-	}
-	
-	// 프로그램 카테고리 수정
-	@GetMapping("/modify-category")
-	public String modifyCategory(@RequestParam(name = "no") int no, 
-			@RequestParam(name = "name") String name) {
-		programService.updateCategory(no, name);
-		
-		return "redirect:category";
-	}
-	
-	// 프로그램 카테고리 삭제
-	@GetMapping("/delete-category")
-	public String deleteCategory(@RequestParam(name = "no") int no, 
-			@RequestParam(name = "name") String name) {
-		programService.deleteCategory(no, name);
-		
-		return "redirect:category";
-	}
-	
-	// 프로그램 카테고리 등록
-	@GetMapping("/insert-category")
-	public String insertCategory(@RequestParam(name = "name") String name) {
-		programService.insertCategory(name);
-		
-		return "redirect:category";
 	}
 }
