+++ src/main/webapp/WEB-INF/views/admin/club/club.jsp
@@ -6,8 +6,8 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous">
 <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.1/font/bootstrap-icons.css">
-<link rel="stylesheet" href="../../../resources/css/common.css">
-<link rel="stylesheet" href="../../../resources/css/content.css">
+<link rel="stylesheet" href="/resources/css/common.css">
+<link rel="stylesheet" href="/resources/css/content.css">
 <title>중앙피트니스</title>
 </head>
 <body>
@@ -73,7 +73,7 @@
 					<sec:authorize access="hasRole('ROLE_ADMIN')">
 						<div class="row">
 							<div class="col-12">
-								<a href="modify?no=${club.no }" class="btn btn-primary float-end">수정</a>
+								<a href="/admin/club/modify?no=${club.no }" class="btn btn-primary float-end">수정</a>
 							</div>
 						</div>
 					</sec:authorize>
