+++ src/main/webapp/WEB-INF/views/admin/club/modify.jsp
@@ -6,8 +6,8 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous">
 <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.1/font/bootstrap-icons.css">
-<link rel="stylesheet" href="../../../resources/css/common.css">
-<link rel="stylesheet" href="../../../resources/css/content.css">
+<link rel="stylesheet" href="/resources/css/common.css">
+<link rel="stylesheet" href="/resources/css/content.css">
 <title>중앙피트니스</title>
 </head>
 <body>
@@ -111,7 +111,7 @@
 					
 						<div class="row">
 							<div class="col-12 text-end">
-								<a href="club" class="btn btn-secondary">취소</a>
+								<a href="/emp/club/club" class="btn btn-secondary">취소</a>
 								<button type="button" id="btn-save-club" class="btn btn-primary">저장</button>
 							</div>
 						</div>
