+++ src/main/webapp/WEB-INF/views/admin/employee/insert.jsp
@@ -6,8 +6,8 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous">
 <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.1/font/bootstrap-icons.css">
-<link rel="stylesheet" href="../../../resources/css/common.css">
-<link rel="stylesheet" href="../../../resources/css/content.css">
+<link rel="stylesheet" href="/resources/css/common.css">
+<link rel="stylesheet" href="/resources/css/content.css">
 <title>중앙피트니스</title>
 </head>
 <body>
@@ -121,7 +121,7 @@
 								</table>
 						
 								<div class="text-end">
-									<a href="list" class="btn btn-secondary">취소</a>
+									<a href="/emp/employee/list" class="btn btn-secondary">취소</a>
 									<button class="btn btn-primary">등록</a>
 								</div>
 							</div>
